soundclient/pyproject.toml
[project]
name = "soundclient"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "realtimestt>=0.3.104",
    "requests>=2.32.5",
]
